Fix loading of popular cards
