+++ src/js/index.js
@@ -1,6 +1,5 @@
 import './teamModal';
 import './searchByKeyword';
-import './renderCards';
 import './cardModal';
 import './pagination';
 import './upBtn';
