+++ src/js/pagination.js
@@ -1,10 +1,16 @@
+//імпорт асинхронної функції феч з отриманням даних з сервера
+import cardTemplate from '../templates/cardTemplate.hbs';
+import NewApiService from './api-servis';
+import {
+  onInputForm,
+  numberOfGeneras,
+  ganreListProcessin,
+} from './searchByKeyword';
 
-//імпорт асинхронної функції феч з отриманням даних з сервера 
-import cardTemplate from '../templates/cardTemplate.hbs'
-import NewApiService from './api-servis'
-import {onInputForm} from './searchByKeyword'
-const newApiService = new NewApiService()
+import NewLoader from './loader';
 
+const newApiService = new NewApiService();
+const newLoader = new NewLoader();
 
 let page = 1;
 let total = 20;
@@ -18,15 +24,15 @@ const pageMainContent = document.querySelector('.main-content');
 const searchForm = document.querySelector('.header__search');
 
 totalAll();
-mainContent();
+ganreListProcessin().then(fetchData);
 renderPagination(page, total);
 addListener();
 switchArrow();
 
 //перемикач по кліку на цифри
 
- function addListener(){
-  const liElItems = document.querySelectorAll(".pagination__btn");
+function addListener() {
+  const liElItems = document.querySelectorAll('.pagination__btn');
 
   for (const liElItem of liElItems) {
     if (liElItem.classList.contains('pagination__points')) {
@@ -36,50 +42,48 @@ switchArrow();
       pageNum(liElItem.textContent * 1);
       newApiService.pageNum = page;
       reseter();
-      mainContent();
+      ganreListProcessin().then(fetchData);
       renderPagination(page, total);
       addListener(liElItems);
 
-      switchArrow ();
-    })
-  }  searchForm.addEventListener('submit', (e) =>{ //перемикач по кліку на цифри, запуск пагінації після пошуку по слову
-      pageNum(liElItem.textContent*1)
-      //newApiService.pageNum = page;
-      reseter();
-      onInputForm();
-      renderPagination(page, total);
-      addListener(liElItems);
-      switchArrow ();
-  }); 
-
+      switchArrow();
+    });
+  }
+  searchForm.addEventListener('submit', e => {
+    //перемикач по кліку на цифри, запуск пагінації після пошуку по слову
+    pageNum(liElItem.textContent * 1);
+    //newApiService.pageNum = page;
+    reseter();
+    onInputForm();
+    renderPagination(page, total);
+    addListener(liElItems);
+    switchArrow();
+  });
 }
 
 //перемикач по кліку на <>
 
-function switchArrow (){
-  const jsBtnArrows = document.querySelectorAll(".js-arrow");
-  
-  for (const jsBtnArrow of jsBtnArrows){
-    jsBtnArrow.addEventListener ('click', (e) => {
-  
-      if (jsBtnArrow.classList.contains("pagination__arrow-left")) {
+function switchArrow() {
+  const jsBtnArrows = document.querySelectorAll('.js-arrow');
 
+  for (const jsBtnArrow of jsBtnArrows) {
+    jsBtnArrow.addEventListener('click', e => {
+      if (jsBtnArrow.classList.contains('pagination__arrow-left')) {
         page = page - 1;
         newApiService.pageNum = page;
         reseter();
-        mainContent();
+        ganreListProcessin().then(fetchData);
         renderPagination(page, total);
         switchArrow();
         addListener();
       }
 
-      
-      if (jsBtnArrow.classList.contains("pagination__arrow-right")) {
-        page = page + 1
+      if (jsBtnArrow.classList.contains('pagination__arrow-right')) {
+        page = page + 1;
 
         newApiService.pageNum = page;
         reseter();
-        mainContent();
+        ganreListProcessin().then(fetchData);
         renderPagination(page, total);
         switchArrow();
         addListener();
@@ -88,33 +92,52 @@ function switchArrow (){
   }
 }
 
-// відбудова карток
-function mainContent() {
-  const pageMainContent = document.querySelector('.main-content');
-  newApiService.fetchPopularMovies().then(({ results }) => {
-    results.map(result => {
-      const { poster_path, original_title, genre_ids, release_date, id } =
-        result;
-      const genres = genre_ids.join(', ');
-      const date = release_date.slice(0, 4);
-      const mk = cardTemplate({
-        poster_path,
-        original_title,
-        genres,
-        date,
-        id,
+// // відбудова карток
+
+async function fetchData(genresBase) {
+  try {
+    newLoader.showLoader();
+    await new Promise(resolve => setTimeout(resolve, 300));
+    await newApiService.fetchPopularMovies().then(({ results }) => {
+      results.map(result => {
+        const { poster_path, original_title, genre_ids, release_date, id } =
+          result;
+        const date = release_date.slice(0, 4);
+
+        const genres = [...genresBase.genres];
+        const genreIds = genre_ids;
+        let genresNames = [];
+        const other = 'Other';
+        for (let i = 0; i < genreIds.length; i++) {
+          const genre = genres.find(g => g.id === genreIds[i]);
+          genresNames.push(genre.name);
+        }
+        let currentGanre = [...genresNames.slice(0, 2), other].join(' ');
+        if (genresNames.length < numberOfGeneras) {
+          currentGanre = [...genresNames].join(' ');
+        }
+
+        const mk = cardTemplate({
+          poster_path,
+          original_title,
+          currentGanre,
+          date,
+          id,
+        });
+
+        const container = document.querySelector('.main-content');
+        container.insertAdjacentHTML('beforeend', mk);
       });
-      const container = document.querySelector('.main-content');
-      container.insertAdjacentHTML('beforeend', mk);
     });
-  });
+    newLoader.hideLoader();
+  } catch (error) {
+    console.error('Помилка під час отримання даних:', error);
+  }
 }
 
 //відбудова кнопок
 
-function renderPagination(page, total) { 
-  
-
+function renderPagination(page, total) {
   if (total < 6) {
     if (page != 1) {
       markup += `<button class="pagination__arrow-left js-arrow">${LEFT_ARROW}</button>`;
@@ -198,8 +221,8 @@ function totalAll() {
   newApiService.fetchPopularMovies().then(res => {
     totalAll = res.total_pages;
 
-    totalPage(totalAll)
-  })
+    totalPage(totalAll);
+  });
 
   function totalPage() {
     if (totalAll > 500) {
